Update backup_v2.sh
the remote name is called "satellite".  When you use this command on the satellite POC, you will need to change the remote name to the correct remote name used in the TigerShark POC.  See the document titled "TigerShark file sharing approach#1_061222_v1.docx, at bottom of the document for the RemoteNames.